change buffer lengths of sysctl(2) to be size_t's
SCCS-vsn: sys/kern/kern_clock.c 7.30
SCCS-vsn: sys/kern/vfs_subr.c 7.95

+++ usr/src/sys/kern/kern_clock.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)kern_clock.c	7.29 (Berkeley) %G%
+ *	@(#)kern_clock.c	7.30 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -460,7 +460,7 @@ statclock(frame)
  */
 sysctl_clockrate(where, sizep)
 	register char *where;
-	int *sizep;
+	size_t *sizep;
 {
 	struct clockinfo clkinfo;
